+++ internal/pkg/ent/price.go
@@ -23,9 +23,9 @@ type Price struct {
 	// EndAt holds the value of the "end_at" field.
 	EndAt time.Time `json:"end_at,omitempty"`
 	// Price holds the value of the "price" field.
-	Price uint16 `json:"price,omitempty"`
+	Price int64 `json:"price,omitempty"`
 	// Discount holds the value of the "discount" field.
-	Discount uint16 `json:"discount,omitempty"`
+	Discount int64 `json:"discount,omitempty"`
 	// Edges holds the relations/edges for other nodes in the graph.
 	// The values are being populated by the PriceQuery when eager-loading is set.
 	Edges                    PriceEdges `json:"edges"`
@@ -116,12 +116,12 @@ func (pr *Price) assignValues(values ...interface{}) error {
 	if value, ok := values[2].(*sql.NullInt64); !ok {
 		return fmt.Errorf("unexpected type %T for field price", values[2])
 	} else if value.Valid {
-		pr.Price = uint16(value.Int64)
+		pr.Price = value.Int64
 	}
 	if value, ok := values[3].(*sql.NullInt64); !ok {
 		return fmt.Errorf("unexpected type %T for field discount", values[3])
 	} else if value.Valid {
-		pr.Discount = uint16(value.Int64)
+		pr.Discount = value.Int64
 	}
 	values = values[4:]
 	if len(values) == len(price.ForeignKeys) {
